Switch to using pyuic5 and pyrcc5 from within Python modules (#1)
Before this update, pyuic5 and pyrcc5 were called externally from Python just like one would for any program to startup. However, there were frequent issues on Windows with whether the shell should be used to call the command. Dipak had an issue where the shell needed to be true, Saygin's old machine required that shell be false and then my machine in the lab was fine with either method.

To get into a bit more detail, I call an external command using `subprocess.run` (same as check_call, check_output, etc but with different arguments). One of the arguments that is useful for this is whether the command should be run in a shell. By not using a shell, it is like you are running the command directly, such as if you did Windows+R on Windows and typed in the command executable. With a shell, a command prompt is first opened (cmd.exe for Windows or /bin/sh for Linux) and then that command is ran from there. There are security issues with turning shell on where vulnerabilities can be exploited by running from the shell, so there are many people advising not to use it. In Sam's case, he is on Debian 9(?) and needed for shell to be off in order for it to run properly. The biggest issue I found with this on Windows is that sometimes subprocess does not search your PATH variable for the executable if shell is off.

As can be seen, it is a pain being able to manage all of this. I'm not entirely sure why two users on Windows got different results with shell set to True or False but it is a mess. Additionally, users must have pyuic5 and pyrcc5 in their path in order for this package to work correctly. This is true even if the user has PyQt5 installed.

**This PR resolves these issues!**

Instead of calling pyuic5 or pyrcc5, rather the command python -m XXX <ARGS> where XXX is the module used for either pyuic5 or pyrcc5. When one calls the command line interfaces for UIC or QRC compilation, they are simply calling a Python module with the specified arguments.

This alone does nothing for the current state of pyqt5ac because now the user must have the correct Python version in their path, which definitely will not always be possible when users have multiple environments for their projects. To fix this, the current Python executable that the script is running will be used to run the commands. This will inherently find the correct pyuic5 and pyrcc5 modules depending on what Python you called the script from.

Tested on my Windows machine and WSL and it seemed to work fine. Need to do more rigorous testing tomorrow.

On an unrelated note, fixed a mistake in the package installation (via GitHub) instructions on the README. I obviously copied and pasted a command and forgot to alter it. I found this mistake when trying to run the command and I was getting a mistake.